+++ simploxide-client/examples/squaring_bot.rs
@@ -15,9 +15,7 @@ use std::{error::Error, sync::Arc};
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
     let (client, mut events) = simploxide_client::connect("ws://127.0.0.1:5225").await?;
-
-    // Use destructuring to extract data from responses
-    let ActiveUserResponse { ref user, .. } = *client.show_active_user().await?;
+    let user = client.show_active_user().await?;
 
     println!(
         "Bot profile: {} ({})",